chore: addressing comments part 1
